Reorganize project structure for better maintainability
- Move development config files to dedicated config/ directory
  - ESLint, Jest, Prettier, TypeScript, Playwright, and Next.js configs
- Move Docker-related files to docker/ directory
- Create symlinks for essential files (docker-compose.yaml, .dockerignore, tsconfig.json)
- Update package.json scripts to reference new file locations
- Add minimal config files in root that reference the moved files

This change improves project organization while maintaining compatibility with all tools and CI processes.
The reorganization makes the root directory cleaner and more manageable, while ensuring that all tools
continue to work as expected by using symlinks and reference files.
