+++ .dockerignore
@@ -1,45 +1,1 @@
-# Dependencies
-node_modules
-.pnp
-.pnp.js
-.pnpm-store
-.node_modules/*
-
-
-# Testing
-coverage
-
-# Next.js
-.next
-out
-
-# Production
-build
-
-# Misc
-.DS_Store
-*.pem
-
-# Debug
-npm-debug.log*
-yarn-debug.log*
-yarn-error.log*
-.pnpm-debug.log*
-
-# Local env files
-.env*.local
-
-# Vercel
-.vercel
-
-# TypeScript
-*.tsbuildinfo
-next-env.d.ts
-
-# Git
-.git
-.gitignore
-
-# Uploaded files - we don't want to include these in the image
-# as they will be mounted as a volume
-public/uploads/*
+docker/.dockerignore
