+++ config/next/next.config.user.mjs
@@ -1,0 +1,55 @@
+// Rename this file to next.config.user.mjs to avoid conflicts with next.config.js
+
+let userConfig = undefined
+try {
+  // try to import ESM first
+  userConfig = await import('./v0-user-next.config.mjs')
+} catch (e) {
+  try {
+    // fallback to CJS import
+    userConfig = await import("./v0-user-next.config");
+  } catch (innerError) {
+    // ignore error
+  }
+}
+
+/** @type {import('next').NextConfig} */
+const nextConfig = {
+  eslint: {
+    ignoreDuringBuilds: false,
+  },
+  typescript: {
+    ignoreBuildErrors: false,
+  },
+  images: {
+    unoptimized: false,
+  },
+  experimental: {
+    webpackBuildWorker: true,
+    parallelServerBuildTraces: true,
+    parallelServerCompiles: true,
+  },
+  // Add output configuration for standalone mode
+  output: 'standalone',
+}
+
+if (userConfig) {
+  // ESM imports will have a "default" property
+  const config = userConfig.default || userConfig
+
+  for (const key in config) {
+    if (
+      typeof nextConfig[key] === 'object' &&
+      !Array.isArray(nextConfig[key])
+    ) {
+      nextConfig[key] = {
+        ...nextConfig[key],
+        ...config[key],
+      }
+    } else {
+      nextConfig[key] = config[key]
+    }
+  }
+}
+
+export default nextConfig
