+++ docker-compose.yaml
@@ -1,42 +1,1 @@
-services:
-  app:
-    image: ghcr.io/kr0osti/image-processor/nextjs-app:${VERSION:-latest}
-    container_name: nextjs-image-processor
-    restart: unless-stopped
-    user: "1001:1001"
-    ports:
-      - "6060:3000"
-    environment:
-      - NODE_ENV=${NODE_ENV:-production}
-      - NODE_OPTIONS=${NODE_OPTIONS:-"--max-old-space-size=4096"}
-      - NEXT_PUBLIC_DEBUG=${NEXT_PUBLIC_DEBUG:-false}
-      - HOSTNAME=${HOSTNAME:-0.0.0.0}
-      - NEXT_PUBLIC_SITE_NAME=${NEXT_PUBLIC_SITE_NAME}
-      - NEXT_PUBLIC_SITE_DESCRIPTION=${NEXT_PUBLIC_SITE_DESCRIPTION}
-      - NEXT_PUBLIC_SITE_URL=${NEXT_PUBLIC_SITE_URL}
-      - NEXT_PUBLIC_SITE_THEME_COLOR=${NEXT_PUBLIC_SITE_THEME_COLOR}
-      - NEXT_PUBLIC_SITE_BACKGROUND_COLOR=${NEXT_PUBLIC_SITE_BACKGROUND_COLOR}
-      - CLEANUP_API_KEY=${CLEANUP_API_KEY}
-    volumes:
-      - ./custom-icons:/app/public/custom-icons
-      - ./public/uploads:/app/public/uploads
-    healthcheck:
-      test: ["CMD-SHELL", "echo 'Debug setting:' $NEXT_PUBLIC_DEBUG && curl -f http://localhost:3000/api/healthcheck || exit 1"]
-      interval: 30s
-      timeout: 10s
-      retries: 3
-      start_period: 30s
-
-  cleanup-cron:
-    image: ghcr.io/kr0osti/image-processor/cleanup-service:${VERSION:-latest}
-    restart: unless-stopped
-    container_name: nextjs-image-cleanup
-    depends_on:
-      - app
-    environment:
-      - CLEANUP_API_KEY=${CLEANUP_API_KEY}
-    volumes:
-      - cleanup-logs:/var/log
-
-volumes:
-  cleanup-logs:
+docker/docker-compose.yaml
