+++ jest.config.js
@@ -1,46 +1,2 @@
-const nextJest = require('next/jest');
-
-const createJestConfig = nextJest({
-  // Provide the path to your Next.js app to load next.config.js and .env files in your test environment
-  dir: './',
-});
-
-// Add any custom config to be passed to Jest
-const customJestConfig = {
-  setupFilesAfterEnv: [
-    process.env.NODE_ENV === 'test-node' ? '<rootDir>/jest.node.setup.js' : '<rootDir>/jest.setup.js',
-  ],
-  testEnvironment: 'jest-environment-jsdom',
-  forceExit: true,
-  testMatch: [
-    '**/__tests__/**/*.test.(js|jsx|ts|tsx)',
-    '**/?(*.)+(spec|test).(js|jsx|ts|tsx)',
-  ],
-  moduleNameMapper: {
-    // Handle module aliases (if you have them in tsconfig.json)
-    '^@/(.*)$': '<rootDir>/$1',
-  },
-  // Add more setup options before each test is run
-  // setupFilesAfterEnv: ['<rootDir>/jest.setup.js'],
-  testPathIgnorePatterns: ['<rootDir>/node_modules/', '<rootDir>/.next/', '<rootDir>/e2e/'],
-  transform: {
-    // Use babel-jest to transpile tests with the next/babel preset
-    '^.+\\.(js|jsx|ts|tsx)$': ['babel-jest', { presets: ['next/babel'] }],
-  },
-  transformIgnorePatterns: [
-    '/node_modules/',
-    '^.+\\.module\\.(css|sass|scss)$',
-  ],
-  collectCoverage: true,
-  collectCoverageFrom: [
-    'app/**/*.{js,jsx,ts,tsx}',
-    'lib/**/*.{js,jsx,ts,tsx}',
-    '!**/*.d.ts',
-    '!**/node_modules/**',
-    '!**/.next/**',
-    '!**/coverage/**',
-  ],
-};
-
-// createJestConfig is exported this way to ensure that next/jest can load the Next.js config which is async
-module.exports = createJestConfig(customJestConfig);
+// Import the Jest config from the config directory
+module.exports = require('./config/jest/jest.config.js');
