+++ package.json
@@ -9,11 +9,11 @@
     "dev": "next dev",
     "build": "next build",
     "start": "next start",
-    "lint": "next lint",
-    "test": "jest",
-    "test:watch": "jest --watch",
-    "test:coverage": "jest --coverage",
-    "test:e2e": "playwright test"
+    "lint": "next lint --config config/eslint/.eslintrc.json",
+    "test": "jest --config config/jest/jest.config.js",
+    "test:watch": "jest --watch --config config/jest/jest.config.js",
+    "test:coverage": "jest --coverage --config config/jest/jest.config.js",
+    "test:e2e": "playwright test --config config/playwright/playwright.config.ts"
   },
   "dependencies": {
     "@hookform/resolvers": "^3.9.1",
